pyproject.toml
[tool.poetry]
name = "django-mini-ecommerce"
version = "0.1.0"
description = "Django Mini Ecommerce System"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
djangorestframework = "^3.13.1"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.2"
autopep8 = "^1.6.0"
djangorestframework-simplejwt = "^5.0.0"
dj-database-url = "^0.5.0"
django-cors-headers = "^3.10.1"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
